Add debug mode button in login and remove in connection screen
Added a "Debug mode" button to the login screen for easier access to debug functionalities and removed the debug button from the connection screen.
Also, added a `DebugScreen` composable file.
Moved the `scheduleApiWorker` method, probably in a better position.
Increased font size of login button and debug button to 18.sp.